[doc] move documentation to asciidoc
If we want to include the plugin documentation in elastic website, we first need to move it to asciidoc format.

By convention we put the documentation in `src/main/asciidoc` as it is expected by some asciidoc maven plugins.
Might make sense to put it in `src/main/docs` though.